multi: apply the new type `lntypes.WeightUnit`

+++ itest/lnd_channel_funding_fund_max_test.go
@@ -309,7 +309,7 @@ func fundingFee(numInput int, change bool) btcutil.Amount {
 		weightEstimate.AddP2TROutput()
 	}
 
-	totalWeight := int64(weightEstimate.Weight())
+	totalWeight := weightEstimate.Weight()
 
 	return feeRate.FeeForWeight(totalWeight)
 }
